+++ kernel/src/fs.rs
@@ -6,3 +6,4 @@
 pub mod elf;
 pub mod fat;
 pub mod tar;
+pub mod vfs;