+++ kernel/src/shell/cmds/move.rs
@@ -3,7 +3,7 @@
 //! Implementation of the file move/rename command.
 
 use crate::io::vga;
-use crate::fs::fat::{read_file_content, create_file, write_file_content, remove_entry};
+use crate::fs::vfs;
 
 static mut MOVE_BUFFER: [u8; 65536] = [0; 65536];
 
@@ -27,15 +27,15 @@ pub fn run(parts: &mut core::str::SplitWhitespace) {
     unsafe {
         let move_buf = &mut *core::ptr::addr_of_mut!(MOVE_BUFFER);
         // Read source file content into the static move buffer
-        match read_file_content(src, move_buf) {
+        match vfs::read_file(src, move_buf) {
             Ok(size) => {
                 // Try to create the destination file (if it already exists, we will overwrite it)
-                let _ = create_file(dest);
+                let _ = vfs::create_file(dest);
 
-                match write_file_content(dest, &move_buf[..size]) {
+                match vfs::write_file(dest, &move_buf[..size]) {
                     Ok(_) => {
                         // Delete the source file after a successful copy
-                        match remove_entry(src) {
+                        match vfs::remove_entry(src) {
                             Ok(_) => {
                                 vga::print_str("File moved successfully.\n");
                             }